+++ web/package.json
@@ -5,27 +5,12 @@
   "dependencies": {
     "@buf/connectrpc_eliza.bufbuild_es": "^2.2.3-20230913231627-233fca715f49.1",
     "@buf/jlewi_foyle.bufbuild_es": "^2.2.3-20241219050144-856e893555b3.1",
-    "@bufbuild/protobuf": "^2.2.3",
-    "@connectrpc/connect": "^2.0.2",
-    "@connectrpc/connect-web": "^2.0.2",
-    "@monaco-editor/react": "^4.7.0",
-    "@types/uuid": "^10.0.0",
     "@buf/stateful_runme.bufbuild_es": "^2.2.3-20250313184526-2d02c6150545.1",
     "@bufbuild/protobuf": "^2.2.3",
+    "@connectrpc/connect": "^2.0.2",
     "@connectrpc/connect-node": "^2.0.2",
     "@connectrpc/connect-web": "^2.0.2",
-    "@connectrpc/connect": "^2.0.2",
     "@monaco-editor/react": "^4.7.0",
-    "@types/uuid": "^10.0.0",
-    "@types/vscode-notebook-renderer": "^1.72.3",        
-    "react-dom": "^18",
-    "react-markdown": "^9.0.1",
-    "react": "^18",
-    "socket.io-client": "^4.8.1",
-    "socket.io": "^4.8.1",    
-    "tsx": "^4.19.3",
-    "ulid": "^2.4.0",
-    "uuid": "^11.1.0",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.2.0",
@@ -34,14 +19,22 @@
     "@types/node": "^16.18.126",
     "@types/react": "^19.0.12",
     "@types/react-dom": "^19.0.4",
+    "@types/uuid": "^10.0.0",
+    "@types/vscode-notebook-renderer": "^1.72.3",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-helmet": "^6.1.0",
+    "react-markdown": "^9.0.1",
     "react-router-dom": "^7.4.0",
     "react-scripts": "5.0.1",
+    "socket.io": "^4.8.1",
+    "socket.io-client": "^4.8.1",
+    "tsx": "^4.19.3",
     "typescript": "^4.9.5",
+    "ulid": "^2.4.0",
+    "uuid": "^11.1.0",
     "web-vitals": "^2.1.4"
   },
-  
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
